requirements.txt
absl-py==1.0.0
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astropy==5.2.1
asttokens==2.0.5
astunparse==1.6.3
attrs==21.4.0
Babel==2.10.1
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
blinker==1.4
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
Click==7.0
colorama==0.4.3
cryptography==2.8
cycler==0.11.0
Cython==0.29.33
dbus-python==1.2.16
debugpy==1.6.0
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distro==1.4.0
distro-info===0.23ubuntu1
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
flatbuffers==2.0
fonttools==4.33.3
gala==0.5.dev0
gast==0.5.3
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.45.0
h5py==3.6.0
httplib2==0.14.0
idna==3.3
imageio==2.26.0
importlib-metadata==4.11.3
importlib-resources==5.7.1
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.6
jsonschema==4.4.0
jupyter-client==7.3.0
jupyter-core==4.10.0
jupyter-server==1.17.0
jupyterlab==3.3.4
jupyterlab-language-pack-zh-CN==3.3.post3
jupyterlab-pygments==0.2.2
jupyterlab-server==2.13.0
keras==2.8.0
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.4.2
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-loader==0.1
libclang==14.0.1
libtiff==0.4.2
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
nbclassic==0.3.7
nbclient==0.6.0
nbconvert==6.5.0
nbformat==5.3.0
nest-asyncio==1.5.5
networkx==3.0
notebook==6.4.11
notebook-shim==0.1.0
numpy==1.22.3
oauthlib==3.2.0
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.20.1
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyerfa==2.0.0.1
Pygments==2.12.0
PyGObject==3.36.0
PyJWT==1.7.1
pyparsing==3.0.8
pyrsistent==0.18.1
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
pytz==2022.1
PyWavelets==1.4.1
PyYAML==5.3.1
pyzmq==22.3.0
requests==2.27.1
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rsa==4.8
scikit-image==0.16.2
scikit-learn==1.2.1
scipy==1.8.0
seaborn==0.11.2
SecretStorage==2.3.1
Send2Trash==1.8.0
simplejson==3.16.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
ssh-import-id==5.10
stack-data==0.2.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-gpu==2.8.0
tensorflow-io-gcs-filesystem==0.25.0
termcolor==1.1.0
terminado==0.13.3
tf-estimator-nightly==2.8.0.dev2021122109
thop==0.0.31.post2005241907
threadpoolctl==3.1.0
tifffile==2023.2.28
tinycss2==1.1.1
torch==1.8.1+cu111
torchvision==0.9.1+cu111
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
typing-extensions==4.2.0
ubuntu-advantage-tools==27.7
ubuntu-drivers-common==0.0.0
unzip==1.0.0
urllib3==1.26.9
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.2
Werkzeug==2.1.2
wrapt==1.14.0
xkit==0.0.0
xlwt==1.3.0
yolk==0.4.3
zipp==3.8.0
